refactor(auth-nest): introduce auth engine seam for legacy jwt flow
Introduce the internal AuthEnginePort boundary in auth-nest and move the
current JWT runtime behavior behind that seam without changing the public auth
HTTP contract.

- add AuthOrchestrationService as the controller-facing AuthService
  implementation
- expand AuthEnginePort with login, logout, and refreshTokens while preserving
  the existing spike capability methods
- move legacy JWT login, logout, refresh-token validation, and token issuance
  into LegacyJwtAuthEngineAdapter
- rewire AuthApplicationModule so AuthService resolves through the
  orchestration layer and AuthEnginePort resolves through the legacy JWT adapter
- keep JwtAuthService exported as a deprecated compatibility wrapper
- keep the Better Auth spike adapter and harness compatible with the merged
  port shape
- add unit coverage for orchestration, legacy JWT adapter, module wiring, and
  compatibility behavior

Validation:
- yarn nx run auth-nest:test
- yarn nx run auth-nest:build
- yarn nx run release-tools:verify-domain-boundaries

Refs #106
Closes #122
Refs #123
