+++ libs/auth/nest/src/application/services/auth-engine.port.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import {
   LoginRequestDTO,
   LoginResponseDTO,
+  LogoutRequestDTO,
   RefreshTokenRequestDTO,
 } from '@anarchitects/auth-ts/dtos';
 
@@ -41,13 +42,16 @@ export type AuthSignOutOrRefreshInput =
 
 @Injectable()
 export abstract class AuthEnginePort {
+  abstract login(dto: LoginRequestDTO): Promise<LoginResponseDTO>;
+  abstract logout(dto: LogoutRequestDTO): Promise<{ success: boolean }>;
+  abstract refreshTokens(
+    userId: string,
+    dto: RefreshTokenRequestDTO,
+  ): Promise<LoginResponseDTO>;
+
   abstract describeCapabilities(): Promise<AuthEngineCapabilityReport>;
-
   abstract passwordSignIn(dto: LoginRequestDTO): Promise<LoginResponseDTO>;
-
   abstract passkeySignIn(input: AuthPasskeySignInInput): Promise<unknown>;
-
   abstract socialSignIn(input: AuthSocialSignInInput): Promise<unknown>;
-
   abstract signOutOrRefresh(input: AuthSignOutOrRefreshInput): Promise<unknown>;
 }
