+++ libs/auth/nest/src/application/services/jwt-auth.service.ts
@@ -1,290 +1,29 @@
-import {
-  ActivateUserRequestDTO,
-  ChangePasswordRequestDTO,
-  ForgotPasswordRequestDTO,
-  LoginRequestDTO,
-  LoginResponseDTO,
-  LogoutRequestDTO,
-  RefreshTokenRequestDTO,
-  RegisterRequestDTO,
-  RegisterResponseDTO,
-  ResetPasswordRequestDTO,
-  UpdateEmailRequestDTO,
-  VerifyEmailRequestDTO,
-} from '@anarchitects/auth-ts/dtos';
-import { PolicyRule, Role, User } from '@anarchitects/auth-ts/models';
-import {
-  BadRequestException,
-  Injectable,
-  InternalServerErrorException,
-} from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
+import { LegacyJwtAuthEngineAdapter } from '../../infrastructure-engine/legacy-jwt-auth-engine.adapter';
 import { AuthUserRepository } from '../../infrastructure-persistence/repositories/auth-user.repository';
-import { AuthService } from './auth.service';
+import { AuthOrchestrationService } from './auth-orchestration.service';
 import { HashService } from './hash.service';
-import { toValidatedPersistedPolicyRule } from './persisted-policy-rule';
 
+/**
+ * @deprecated Use `AuthService` for controller-facing auth operations.
+ * JWT engine access is now an internal wiring concern behind `AuthEnginePort`.
+ */
 @Injectable()
-export class JwtAuthService implements AuthService {
+export class JwtAuthService extends AuthOrchestrationService {
   constructor(
-    private readonly hashService: HashService,
-    private readonly authUserRepository: AuthUserRepository,
-    private readonly jwtService: JwtService,
-  ) {}
-
-  async registerUser(dto: RegisterRequestDTO): Promise<RegisterResponseDTO> {
-    if (dto.password !== dto.confirmPassword) {
-      throw new BadRequestException('Passwords do not match');
-    }
-    const passwordHash = await this.hashService.hash(dto.password);
-    const token = crypto.randomUUID();
-    const { userName, email } = dto;
-    const user: Partial<User> = {
-      email,
-      passwordHash,
-      isActive: false,
-      userName,
-      token,
-      roles: [
-        {
-          name: 'user',
-        } as Role,
-      ],
-    };
-    await this.authUserRepository.create(user);
-    return { success: true };
-  }
-
-  async activateUser(
-    dto: ActivateUserRequestDTO,
-  ): Promise<{ success: boolean }> {
-    const { token } = dto;
-    const user = await this.authUserRepository.findOne(token);
-    await this.authUserRepository.update({
-      ...user,
-      isActive: true,
-      token: null,
-    });
-    return { success: true };
-  }
-
-  async login(dto: LoginRequestDTO): Promise<LoginResponseDTO> {
-    const { credential, password } = dto;
-    const user = await this.authUserRepository.findOne({
-      where: [{ email: credential }, { userName: credential }],
-    });
-    if (!user) {
-      throw new BadRequestException('Invalid credentials');
-    }
-    const isPasswordValid = await this.hashService.compare(
-      password,
-      user.passwordHash,
+    hashService: HashService,
+    authUserRepository: AuthUserRepository,
+    jwtService: JwtService,
+  ) {
+    super(
+      hashService,
+      authUserRepository,
+      new LegacyJwtAuthEngineAdapter(
+        hashService,
+        authUserRepository,
+        jwtService,
+      ),
     );
-    if (!isPasswordValid) {
-      throw new BadRequestException('Invalid credentials');
-    }
-    return this.generateTokens(user);
-  }
-
-  async logout(dto: LogoutRequestDTO): Promise<{ success: boolean }> {
-    const { accessToken, refreshToken } = dto;
-
-    if (!refreshToken) {
-      throw new BadRequestException('Refresh token is required');
-    }
-
-    const payload = await this.jwtService
-      .verifyAsync(refreshToken)
-      .catch(() => {
-        throw new BadRequestException('Invalid refresh token');
-      });
-
-    if (!payload?.sub) {
-      throw new BadRequestException('Invalid refresh token');
-    }
-
-    const user = await this.authUserRepository.findOne({
-      where: { id: payload.sub },
-    });
-
-    if (!user?.id) {
-      throw new BadRequestException('Invalid refresh token');
-    }
-
-    const tokenHashes = await Promise.all(
-      [accessToken, refreshToken]
-        .filter((token): token is string => Boolean(token))
-        .map((token) => this.hashService.hash(token)),
-    );
-
-    await this.authUserRepository.invalidateTokens(tokenHashes, user.id);
-
-    return { success: true };
-  }
-
-  async changePassword(
-    userId: string,
-    dto: ChangePasswordRequestDTO,
-  ): Promise<{ success: boolean }> {
-    const { currentPassword, newPassword, confirmPassword } = dto;
-    if (newPassword !== confirmPassword) {
-      throw new BadRequestException('Passwords do not match');
-    }
-    const user = await this.authUserRepository.findOne({
-      where: { id: userId },
-    });
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-    const isCurrentPasswordValid = await this.hashService.compare(
-      currentPassword,
-      user.passwordHash,
-    );
-    if (!isCurrentPasswordValid) {
-      throw new BadRequestException('Invalid current password');
-    }
-    user.passwordHash = await this.hashService.hash(newPassword);
-    await this.authUserRepository.update(user);
-    return { success: true };
-  }
-  async forgotPassword(
-    dto: ForgotPasswordRequestDTO,
-  ): Promise<{ success: boolean }> {
-    const { email } = dto;
-    const user = await this.authUserRepository.findOne({ where: { email } });
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-    const token = crypto.randomUUID();
-    user.token = token;
-    await this.authUserRepository.update(user);
-    // Here you would typically send the token to the user's email
-    return { success: true };
-  }
-  async resetPassword(
-    dto: ResetPasswordRequestDTO,
-  ): Promise<{ success: boolean }> {
-    const { token, password, confirmPassword } = dto;
-    if (password !== confirmPassword) {
-      throw new BadRequestException('Passwords do not match');
-    }
-    const user = await this.authUserRepository.findOne({ where: { token } });
-    if (!user) {
-      throw new BadRequestException('Invalid token');
-    }
-    user.passwordHash = await this.hashService.hash(password);
-    user.token = null;
-    await this.authUserRepository.update(user);
-    return { success: true };
-  }
-  async verifyEmail(dto: VerifyEmailRequestDTO): Promise<{ success: boolean }> {
-    const { token } = dto;
-    const user = await this.authUserRepository.findOne({ where: { token } });
-    if (!user) {
-      throw new BadRequestException('Invalid token');
-    }
-    user.isActive = true;
-    user.token = null;
-    await this.authUserRepository.update(user);
-    return { success: true };
-  }
-
-  async updateEmail(
-    userId: string,
-    dto: UpdateEmailRequestDTO,
-  ): Promise<{ success: boolean }> {
-    const { newEmail, password } = dto;
-    const user = await this.authUserRepository.findOne({
-      where: { id: userId },
-    });
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-    const isPasswordValid =
-      password && (await this.hashService.compare(password, user.passwordHash));
-    if (!isPasswordValid) {
-      throw new BadRequestException('Invalid password');
-    }
-    user.email = newEmail;
-    await this.authUserRepository.update(user);
-    return { success: true };
-  }
-
-  async refreshTokens(
-    userId: string,
-    dto: RefreshTokenRequestDTO,
-  ): Promise<LoginResponseDTO> {
-    const { refreshToken } = dto;
-    const payload = await this.jwtService
-      .verifyAsync(refreshToken)
-      .catch(() => {
-        throw new BadRequestException('Invalid refresh token');
-      });
-
-    if (!payload?.sub || payload.sub !== userId) {
-      throw new BadRequestException('Invalid refresh token');
-    }
-
-    const user = await this.authUserRepository.findOne({
-      where: { id: userId },
-    });
-
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-
-    const isTokenInvalidated = await this.authUserRepository.isTokenInvalidated(
-      await this.hashService.hash(refreshToken),
-    );
-
-    if (isTokenInvalidated) {
-      throw new BadRequestException('Refresh token has been invalidated');
-    }
-
-    return this.generateTokens(user);
-  }
-
-  async getLoggedInUserInfo(
-    userId: string,
-  ): Promise<{ user: User; rbac: PolicyRule[] }> {
-    const user = await this.authUserRepository.findOne({
-      where: { id: userId },
-      relations: ['roles', 'roles.permissions'],
-    });
-    if (!user) {
-      throw new BadRequestException('User not found');
-    }
-
-    const rbac = this.getValidatedPolicyRules(user);
-    return { user, rbac };
-  }
-
-  private getValidatedPolicyRules(user: User): PolicyRule[] {
-    try {
-      const rbac: PolicyRule[] = [];
-      user.roles?.forEach((role) => {
-        role.permissions?.forEach((permission) => {
-          rbac.push(toValidatedPersistedPolicyRule(permission));
-        });
-      });
-
-      return rbac;
-    } catch (error) {
-      throw new InternalServerErrorException(
-        `Malformed persisted policy rule payload: ${
-          error instanceof Error ? error.message : String(error)
-        }`,
-      );
-    }
-  }
-
-  private async generateTokens(user: User) {
-    const payload = {
-      sub: user.id,
-      roles: user.roles?.map((role) => role.name),
-    };
-    const accessToken = await this.jwtService.signAsync(payload);
-    const refreshToken = await this.jwtService.signAsync(payload);
-    return { accessToken, refreshToken };
   }
 }
